refactor(wl): narrow contexts and split frame activity
Replace direct compositor context state access with named capability methods and split frame-loop redraw policy out of the frame side-effect path.

Context changes:
- Restrict mutable context state fields to compositor internals instead of crate-wide access.
- Add capability methods for spawn reveal sizing and activation, fullscreen enter and exit, layer-shell registration/removal/focus/popup constraints, pointer cursor position hints, and surface lifecycle operations.
- Keep the broad toplevel-destroy lifecycle behavior encapsulated inside SurfaceLifecycleCtx for a later lifecycle split, while removing direct ctx.st usage from call sites.

Frame-loop changes:
- Add frame_loop::activity for read-only per-output animation/redraw state and full-repaint decisions.
- Keep frame ticking, frame callbacks, and presentation feedback in frame_loop.rs.
- Preserve existing repaint behavior while separating redraw policy from frame side effects.

Verification:
- cargo fmt --all
- cargo check -p halley-wl
- cargo test -p halley-wl
