+++ crates/halley-wl/src/compositor/workspace/lifecycle/mod.rs
@@ -31,12 +31,172 @@ use surface::{
 
 const CLUSTER_OVERFLOW_PROMOTION_ANIM_MS: u64 = 360;
 
+impl SurfaceLifecycleCtx<'_> {
+    pub(crate) fn refresh_node_identity_for_surface(
+        &mut self,
+        surface: &WlSurface,
+        fallback_label: &str,
+    ) {
+        refresh_node_identity_for_surface(self.st, surface, fallback_label);
+    }
+
+    pub(crate) fn note_commit(&mut self, surface: &WlSurface, now: Instant) {
+        note_commit(self.st, surface, now);
+    }
+
+    pub(crate) fn ensure_node_for_surface(
+        &mut self,
+        surface: &WlSurface,
+        label: &str,
+        size_px: (i32, i32),
+        intent: &InitialWindowIntent,
+    ) -> NodeId {
+        ensure_node_for_surface_impl(self.st, surface, label, size_px, intent)
+    }
+
+    pub(crate) fn drop_surface(&mut self, surface: &WlSurface) {
+        drop_surface_impl(self.st, surface);
+    }
+
+    pub(crate) fn on_toplevel_destroyed(&mut self, surface: ToplevelSurface) {
+        let st = &mut self.st;
+        let key = surface.wl_surface().id();
+        let closing_id = st.model.surface_to_node.get(&key).copied();
+        let closing_fullscreen = closing_id.is_some_and(|id| st.is_fullscreen_active(id));
+        let had_keyboard_focus = st
+            .platform
+            .seat
+            .get_keyboard()
+            .and_then(|kb| kb.current_focus())
+            .is_some_and(|focused| surface_tree_root(&focused).id() == key);
+        let had_pointer_focus = st
+            .platform
+            .seat
+            .get_pointer()
+            .and_then(|ptr| ptr.current_focus())
+            .is_some_and(|focused| surface_tree_root(&focused).id() == key);
+        let focused_monitor = st
+            .model
+            .surface_to_node
+            .get(&key)
+            .and_then(|id| st.model.monitor_state.node_monitor.get(id))
+            .cloned();
+
+        if had_keyboard_focus || had_pointer_focus {
+            eventline::debug!(
+                "toplevel_destroyed with active focus (keyboard={} pointer={}); scheduling input state reset",
+                had_keyboard_focus,
+                had_pointer_focus
+            );
+            st.input.interaction_state.reset_input_state_requested = true;
+            if let Some(ref focused_monitor) = focused_monitor {
+                st.model.spawn_state.pending_spawn_monitor = Some(focused_monitor.clone());
+                eventline::debug!(
+                    "pending spawn monitor latched from destroyed toplevel: {}",
+                    focused_monitor
+                );
+            }
+        }
+
+        if had_keyboard_focus {
+            st.clear_keyboard_focus();
+        }
+
+        let mut deferred_close_restore = None;
+        if had_keyboard_focus
+            && st.runtime.tuning.close_restore_focus
+            && let (Some(closing_id), Some(focused_monitor)) =
+                (closing_id, focused_monitor.as_deref())
+        {
+            let now = Instant::now();
+            let suppress_restore_pan =
+                st.node_has_overlap_policy(closing_id) || st.is_fullscreen_active(closing_id);
+            if closing_fullscreen {
+                if let Some(target) =
+                    non_cluster_close_restore_target(st, focused_monitor, closing_id)
+                {
+                    deferred_close_restore = Some((
+                        focused_monitor.to_string(),
+                        target,
+                        suppress_restore_pan,
+                        now,
+                    ));
+                }
+            } else if let Some(cid) = st.active_cluster_workspace_for_monitor(focused_monitor) {
+                if matches!(
+                    st.runtime.tuning.cluster_layout_kind(),
+                    halley_core::cluster_layout::ClusterWorkspaceLayoutKind::Tiling
+                ) {
+                    // Tiled cluster close restore is handled after the member is actually removed so
+                    // we can focus the replacement tile in that slot.
+                } else {
+                    let mut next_to_focus = None;
+                    if let Some(cluster) = st.model.field.cluster(cid) {
+                        let members = cluster.members();
+                        if let Some(pos) = members.iter().position(|&id| id == closing_id) {
+                            if pos + 1 < members.len() {
+                                next_to_focus = Some(members[pos + 1]);
+                            } else if pos > 0 {
+                                next_to_focus = Some(members[pos - 1]);
+                            }
+                        }
+                    }
+
+                    if let Some(next) = next_to_focus {
+                        st.set_interaction_focus(Some(next), 30_000, now);
+                    } else if let Some(previous) =
+                        st.previous_window_from_trail_on_close(focused_monitor, closing_id)
+                    {
+                        st.set_interaction_focus(Some(previous), 30_000, now);
+                    } else if let Some(fallback) = st
+                        .last_focused_surface_node_for_monitor(focused_monitor)
+                        .filter(|&id| id != closing_id)
+                    {
+                        st.set_interaction_focus(Some(fallback), 30_000, now);
+                    }
+                }
+            } else if let Some(target) =
+                non_cluster_close_restore_target(st, focused_monitor, closing_id)
+            {
+                let _ = st.restore_focus_to_node_after_close(
+                    focused_monitor,
+                    target,
+                    now,
+                    suppress_restore_pan,
+                );
+            }
+        } else if had_keyboard_focus
+            && !st.runtime.tuning.close_restore_focus
+            && let Some(focused_monitor) = focused_monitor.as_deref()
+        {
+            st.model
+                .focus_state
+                .blocked_monitor_focus_restore
+                .insert(focused_monitor.to_string());
+        }
+        if had_pointer_focus {
+            crate::compositor::interaction::pointer::clear_pointer_focus(st);
+        }
+
+        drop_surface_impl(st, surface.wl_surface());
+
+        if let Some((monitor, target, suppress_restore_pan, now)) = deferred_close_restore {
+            let _ = st.restore_focus_to_node_after_close(
+                monitor.as_str(),
+                target,
+                now,
+                suppress_restore_pan,
+            );
+        }
+    }
+}
+
 pub(crate) fn refresh_surface_identity(
     ctx: &mut SurfaceLifecycleCtx<'_>,
     surface: &WlSurface,
     fallback_label: &str,
 ) {
-    refresh_node_identity_for_surface(ctx.st, surface, fallback_label);
+    ctx.refresh_node_identity_for_surface(surface, fallback_label);
 }
 
 pub(crate) fn on_surface_commit(
@@ -44,7 +204,7 @@ pub(crate) fn on_surface_commit(
     surface: &WlSurface,
     now: Instant,
 ) {
-    note_commit(ctx.st, surface, now);
+    ctx.note_commit(surface, now);
 }
 
 pub(crate) fn ensure_node_for_surface(
@@ -54,142 +214,16 @@ pub(crate) fn ensure_node_for_surface(
     size_px: (i32, i32),
     intent: &InitialWindowIntent,
 ) -> NodeId {
-    ensure_node_for_surface_impl(ctx.st, surface, label, size_px, intent)
+    ctx.ensure_node_for_surface(surface, label, size_px, intent)
 }
 
 #[allow(dead_code)]
 pub(crate) fn drop_surface(ctx: &mut SurfaceLifecycleCtx<'_>, surface: &WlSurface) {
-    drop_surface_impl(ctx.st, surface);
+    ctx.drop_surface(surface);
 }
 
 pub(crate) fn on_toplevel_destroyed(ctx: &mut SurfaceLifecycleCtx<'_>, surface: ToplevelSurface) {
-    let st = &mut ctx.st;
-    let key = surface.wl_surface().id();
-    let closing_id = st.model.surface_to_node.get(&key).copied();
-    let closing_fullscreen = closing_id.is_some_and(|id| st.is_fullscreen_active(id));
-    let had_keyboard_focus = st
-        .platform
-        .seat
-        .get_keyboard()
-        .and_then(|kb| kb.current_focus())
-        .is_some_and(|focused| surface_tree_root(&focused).id() == key);
-    let had_pointer_focus = st
-        .platform
-        .seat
-        .get_pointer()
-        .and_then(|ptr| ptr.current_focus())
-        .is_some_and(|focused| surface_tree_root(&focused).id() == key);
-    let focused_monitor = st
-        .model
-        .surface_to_node
-        .get(&key)
-        .and_then(|id| st.model.monitor_state.node_monitor.get(id))
-        .cloned();
-
-    if had_keyboard_focus || had_pointer_focus {
-        eventline::debug!(
-            "toplevel_destroyed with active focus (keyboard={} pointer={}); scheduling input state reset",
-            had_keyboard_focus,
-            had_pointer_focus
-        );
-        st.input.interaction_state.reset_input_state_requested = true;
-        if let Some(ref focused_monitor) = focused_monitor {
-            st.model.spawn_state.pending_spawn_monitor = Some(focused_monitor.clone());
-            eventline::debug!(
-                "pending spawn monitor latched from destroyed toplevel: {}",
-                focused_monitor
-            );
-        }
-    }
-
-    if had_keyboard_focus {
-        st.clear_keyboard_focus();
-    }
-
-    let mut deferred_close_restore = None;
-    if had_keyboard_focus
-        && st.runtime.tuning.close_restore_focus
-        && let (Some(closing_id), Some(focused_monitor)) = (closing_id, focused_monitor.as_deref())
-    {
-        let now = Instant::now();
-        let suppress_restore_pan =
-            st.node_has_overlap_policy(closing_id) || st.is_fullscreen_active(closing_id);
-        if closing_fullscreen {
-            if let Some(target) = non_cluster_close_restore_target(st, focused_monitor, closing_id)
-            {
-                deferred_close_restore = Some((
-                    focused_monitor.to_string(),
-                    target,
-                    suppress_restore_pan,
-                    now,
-                ));
-            }
-        } else if let Some(cid) = st.active_cluster_workspace_for_monitor(focused_monitor) {
-            if matches!(
-                st.runtime.tuning.cluster_layout_kind(),
-                halley_core::cluster_layout::ClusterWorkspaceLayoutKind::Tiling
-            ) {
-                // Tiled cluster close restore is handled after the member is actually removed so
-                // we can focus the replacement tile in that slot.
-            } else {
-                let mut next_to_focus = None;
-                if let Some(cluster) = st.model.field.cluster(cid) {
-                    let members = cluster.members();
-                    if let Some(pos) = members.iter().position(|&id| id == closing_id) {
-                        if pos + 1 < members.len() {
-                            next_to_focus = Some(members[pos + 1]);
-                        } else if pos > 0 {
-                            next_to_focus = Some(members[pos - 1]);
-                        }
-                    }
-                }
-
-                if let Some(next) = next_to_focus {
-                    st.set_interaction_focus(Some(next), 30_000, now);
-                } else if let Some(previous) =
-                    st.previous_window_from_trail_on_close(focused_monitor, closing_id)
-                {
-                    st.set_interaction_focus(Some(previous), 30_000, now);
-                } else if let Some(fallback) = st
-                    .last_focused_surface_node_for_monitor(focused_monitor)
-                    .filter(|&id| id != closing_id)
-                {
-                    st.set_interaction_focus(Some(fallback), 30_000, now);
-                }
-            }
-        } else if let Some(target) =
-            non_cluster_close_restore_target(st, focused_monitor, closing_id)
-        {
-            let _ = st.restore_focus_to_node_after_close(
-                focused_monitor,
-                target,
-                now,
-                suppress_restore_pan,
-            );
-        }
-    } else if had_keyboard_focus
-        && !st.runtime.tuning.close_restore_focus
-        && let Some(focused_monitor) = focused_monitor.as_deref()
-    {
-        st.model
-            .focus_state
-            .blocked_monitor_focus_restore
-            .insert(focused_monitor.to_string());
-    }
-    if had_pointer_focus {
-        crate::compositor::interaction::pointer::clear_pointer_focus(st);
-    }
-
-    drop_surface_impl(st, surface.wl_surface());
-
-    if let Some((monitor, target, suppress_restore_pan, now)) = deferred_close_restore {
-        let _ = st.restore_focus_to_node_after_close(
-            monitor.as_str(),
-            target,
-            now,
-            suppress_restore_pan,
-        );
-    }
+    ctx.on_toplevel_destroyed(surface);
 }
 
 fn non_cluster_close_restore_target(
